Move common funtions to common.py
